feat: remove old bs carousel
