+++ src/js/bootstrap-italia.entry.js
@@ -35,7 +35,6 @@ import {
   NavBarCollapsible,
   Accordion,
   NavScroll,
-  CarouselBI,
   FormValidate,
   ValidatorSelectAutocomplete,
   Input,
@@ -77,7 +76,6 @@ export default {
   AcceptOverlay,
   Accordion,
   BackToTop,
-  CarouselBI,
   Cookiebar,
   Dimmer,
   FormValidate,
